[finishes #163177427] added tests and commetnts to application
