+++ app/api/v1/models/meetup_models.py
@@ -1,6 +1,7 @@
 meetups = {}
 
 class MeetupInfo:
+    """ Meetup information from user input """
     def __init__(self, id, location, topic, happeningOn, tags, images=None):
         self.id = id
         self.location = location
@@ -11,6 +12,7 @@ class MeetupInfo:
         self.number_of_users = len(meetups) + 1   
 
     def add_meetup(self):
+        """ takes the captured information and appends them to meetups db """
         db={
             "id":self.number_of_users,
             "location":self.location,
@@ -19,11 +21,13 @@ class MeetupInfo:
             "happeningOn":self.happeningOn,
             "tags":self.tags
         }
-   
+        #Append the db dictionary to meetups
         meetups.update({self.number_of_users:db})
         return meetups
+    #static methods have no relationship with the class state
     @staticmethod
     def get_meetup(id):
+        """ Check if meetup is present """
         for key_id, value in meetups.items():
             if key_id == id:
                 return meetups[id]
@@ -31,4 +35,5 @@ class MeetupInfo:
 
     @staticmethod
     def get_all_meetups():
+        """ Gets all meetups """
         return(meetups)
