+++ app/api/v1/models/reservation_models.py
@@ -1,4 +1,5 @@
 class Reservation:
+    """ Defines reservations status """
     def __init__(self, meetup_id, topic, status):
         self.meetup_id = meetup_id
         self.topic = topic
@@ -9,5 +10,6 @@ class Reservation:
         }
 
     def meetup_status(self):
+        """ returns the status of the meetup """
         self.db["status"] = self.status        
         return self.db
