+++ app/api/v1/views/meetup_views.py
@@ -1,13 +1,17 @@
-from flask import Blueprint, Flask, jsonify, request, redirect
+from flask import Blueprint, Flask, jsonify, request
 from app.api.v1.models.meetup_models import MeetupInfo
 from app.validators.shared_validators import check_fields
 
+#set up meetup views blueprints
 mtp = Blueprint('meetup_api', __name__)
 
 class MeetupViews:
+    """ Defines the meetup route """
     @mtp.route('/v1/add_meetups', methods = ['POST'])
     def create_meetup():
+        """ fetch the posted information from the user """
         meetup = request.get_json("meetups")
+        #validate user information
         validate_info = ['location','images','topic',
                                 'happeningOn','tags']
         error = check_fields(meetup, validate_info)
@@ -19,6 +23,7 @@ class MeetupViews:
         topic = meetup['topic']
         happeningOn = meetup['happeningOn']
         tags = meetup['tags']
+        #Meetup information is sent to database
         meetup_object = MeetupInfo('',location,topic,happeningOn,
                                             tags,images)
         meetups = meetup_object.add_meetup()
@@ -31,6 +36,7 @@ class MeetupViews:
 
     @mtp.route('/v1/meetups/<int:meetup_id>', methods = ['GET'])
     def get_meetup(meetup_id):
+        """ Gets  specific meetup id """
         meetup = MeetupInfo
         get_meetup = meetup.get_meetup(meetup_id)
         if get_meetup == {}:
@@ -46,6 +52,7 @@ class MeetupViews:
 
     @mtp.route('/v1/get_meetups', methods = ['GET'])
     def get_meetups():
+        """ gets all meetups """
         meetup = MeetupInfo
         fetch_meetups = meetup.get_all_meetups()
         if len(fetch_meetups) == 0:
