+++ app/api/v1/views/question_views.py
@@ -3,11 +3,14 @@ from app.api.v1.models.question_models import AddQuestion
 from app.api.v1.models.meetup_models import MeetupInfo
 from app.validators.shared_validators import check_fields
 
+#set up question views blueprints
 qsn = Blueprint('questions_api', __name__)
 
 class QuestionsVIews:
+    """ Defines the question route """
     @qsn.route('/v1/<int:meetup_id>/post_question', methods = ['POST'])
     def post_question(meetup_id):
+        """ fetch the posted information from the user input """
         meetup = MeetupInfo
         get_meetup = meetup.get_meetup(meetup_id) 
         if get_meetup == {}:
@@ -15,6 +18,7 @@ class QuestionsVIews:
                 "message":"meetup not found"
             }), 404
         question = request.get_json("question")
+        #validate obtained information
         validate_info = ['title','body']
         error = check_fields(question, validate_info)
         if len(error) > 0:
