+++ app/api/v1/views/vote.py
@@ -1,13 +1,16 @@
 from flask import Blueprint, Flask, jsonify
 from app.api.v1.models.question_models import AddQuestion
 
+#set up vote views blueprints
 vt = Blueprint('vote_api', __name__)
 
 class VoteOnQuestions:
-    
+    """ Defines the upvote route """
     @vt.route('/v1/<int:question_id>/upvote', methods = ['PATCH'])
     def upvote_question(question_id):
+        """ defines question upvote """
         question = AddQuestion.get_question(question_id)
+        #handle validation
         if question == {}:
             return jsonify({"message":"question not found"}), 404
         votes = question['votes'] + 1
@@ -21,6 +24,7 @@ class VoteOnQuestions:
     
     @vt.route('/v1/<int:question_id>/downvote', methods = ['PATCH'])
     def down_question(question_id):
+        """ Defines the downvote route """
         question = AddQuestion.get_question(question_id)
         if question == {}:
             return jsonify({"message":"question not found"}), 404
