+++ app/tests/v1/basetest.py
@@ -1,0 +1,107 @@
+import json
+import unittest #this is a unit testing framework
+from app import create_app
+
+
+
+class TestApplication(unittest.TestCase):
+    def setUp(self):
+        self.app = create_app("testing")
+        #set up the test client
+        self.client = self.app.test_client()
+        """ set up user information test data """
+        self.users = {
+                "firstname":"bill",
+                "lastname":"wekesa",
+                "othername":"adams",
+                "username":"wekebill",
+                "email":"ill@wek.com",
+                "phoneNumber":"526",
+                "password":"Tintinabu12"
+            }
+        
+        self.users_two = {
+                "firstname":"bill",
+                "lastname":"wekesa",
+                "othername":"adams",
+                "username":"wekebill",
+                "email":"ill@wek.com",
+                "phoneNumber":"526",
+                "password":"Tintinabu12"
+        }
+        self.users_three = {
+                "firstname":"bill",
+                "lastname":"wekesa",
+                "othername":"adams",
+                "username":"wekesabill",
+                "email":"ill@wek.com",
+                "phoneNumber":"526",
+                "password":"Tintinabu12"
+        }
+        self.users_four = {
+                "firstname":"bill",
+                "lastname":"wekesa",
+                "othername":"adams",
+                "username":"wekbill",
+                "email":"il@wek.com",
+                "phoneNumber":"526",
+                "password":"tintinabu"
+        }
+        self.users_five = {
+                "firstname":"bill",
+                "lastname":"wekesa",
+                "othername":"adams",
+                "username":"wekbill",
+                "email":"ilwek.com",
+                "phoneNumber":"526",
+                "password":"Tintinabu12"
+        }
+
+        """ set up meetup information test data """
+        self.meetups = {
+                "location":"online",
+                "images":"youtube.com/images",
+                "topic":"javascript",
+                "happeningOn":"12/12/2030",
+                "tags":"classes"
+            }
+        self.meetups_one = {
+                "location":"",
+                "images":"youtube.com/images",
+                "topic":"javascript",
+                "happeningOn":"12/12/2030",
+                "tags":"classes"
+            }
+        self.meetups_two = {
+                "images":"youtube.com/images",
+                "topic":"javascript",
+                "happeningOn":"12/12/2030",
+                "tags":"classes"
+            }
+        self.reservation = { 
+                "status":"YES"
+        }
+
+        """ set up questions information test data """
+        self.questions = {
+                "title":"adams",
+                "body":"ill@wek"
+            }
+        self.question_one = {
+                "title":"",
+                "body":"ill@wek"
+        }
+        self.upvote = {
+                "body": "ill@wek",
+                "question": 1,
+                "title": "adams",
+                "votes": 1
+            }
+        
+    def tearDown(self):
+        """ Pull down data to original after testing """
+        self.app.testing = False
+        self.app = None
+
+    if __name__ == '__main__':
+    unittest.main()
