+++ app/validators/shared_validators.py
@@ -1,4 +1,5 @@
 def check_fields(form_data, data_keys):
+    """ Validates that all fields have been provided with data """
     keylist = []
     error_message = {}
     if not form_data:
@@ -8,9 +9,11 @@ def check_fields(form_data, data_keys):
             keylist.append(input_key)
         
         for key_item in data_keys:
+            """ Confirms all field are provided """
             if key_item not in keylist:
                 error_message['key error'] = "{} is missing".format(key_item)
             else:
                 if form_data[key_item].strip() == "":
+                    """ Confirms all fields have values """
                     error_message['value error'] = "{} is missing a value".format(key_item)
     return error_message
