+++ run.py
@@ -1,7 +1,11 @@
+""" os provides a way of using functionalities dependent on the operating sysytem """
 import os
 from app import create_app
 
+""" Return the value of the environment variable """
 config_name = os.getenv('flask_config')
+
+""" Environment variables are sent to create_app and gets configurations to run the app """
 app = create_app(config_name)
 
 if __name__ == '__main__':
